+++ packages/ripple-keypairs/src/utils.ts
@@ -2,18 +2,19 @@ import * as assert from 'assert'
 import * as hashjs from 'hash.js'
 import * as BN from 'bn.js'
 
-function bytesToHex(a): string {
-  return a
-    .map((byteValue) => {
-      const hex = byteValue.toString(16).toUpperCase()
-      return hex.length > 1 ? hex : `0${hex}`
-    })
-    .join('')
+function bytesToHex(a: Iterable<number> | ArrayLike<number>): string {
+  return Array.from(a, (byteValue) => {
+    const hex = byteValue.toString(16).toUpperCase()
+    return hex.length > 1 ? hex : `0${hex}`
+  }).join('')
 }
 
 function hexToBytes(a): number[] {
   assert.ok(a.length % 2 === 0)
-  return new BN(a, 16).toArray(null, a.length / 2)
+  // Special-case length zero to return [].
+  // BN.toArray intentionally returns [0] rather than [] for length zero,
+  // which may make sense for BigNum data, but not for byte strings.
+  return a.length === 0 ? [] : new BN(a, 16).toArray(null, a.length / 2)
 }
 
 function computePublicKeyHash(publicKeyBytes: Buffer): Buffer {