+++ packages/ripple-keypairs/HISTORY.md
@@ -1,5 +1,10 @@
 # ripple-keypairs Release History
 
+## Unreleased
+- `hexToBytes` now produces empty output for empty input, rather than `[0]`.
+- Extend `bytesToHex` to work correctly with any input type accepted by `Array.from`.
+  In particular, it now produces correct output for typed arrays such as `UInt8Array`.
+
 ## 1.1.1 (2021-12-1)
 - Fix issue where npm < 7 was not allowed to install the library